regex impl

+++ src/state.rs
@@ -15,7 +15,7 @@ impl Traversable for char {
 }
 
 pub trait State { 
-    fn next(&self, str: &str) -> Vec<(i32, String)>;
+    fn next(&self, str: &str) -> Vec<(usize, String)>;
 }
 
 #[cfg(test)]
